+++ packages/cli/src/lib/generateBlockData.ts
@@ -242,9 +242,11 @@ function getFromContext(
  * @param fullPath The path to the .appsemblerc file
  */
 export default function generateBlockData(config: BlockConfig, fullPath: string): BlockPayload {
-  const { layout, resources, version } = config;
+  const { description, layout, name, resources, version } = config;
   const { actions, events, parameters } = getFromContext(config, fullPath);
 
+  logger.verbose(`Using name: ${inspect(name, { colors: true, depth: 20 })}`);
+  logger.verbose(`Using description: ${inspect(description, { colors: true, depth: 20 })}`);
   logger.verbose(`Using version: ${inspect(version, { colors: true, depth: 20 })}`);
   logger.verbose(`Using layout: ${inspect(layout, { colors: true, depth: 20 })}`);
   logger.verbose(`Using actions: ${inspect(actions, { colors: true, depth: 20 })}`);
@@ -252,6 +254,8 @@ export default function generateBlockData(config: BlockConfig, fullPath: string)
   logger.verbose(`Using parameters: ${inspect(parameters, { colors: true, depth: 20 })}`);
 
   return {
+    name,
+    description,
     actions,
     events,
     layout,