+++ packages/server/src/controllers/blocks.test.js
@@ -1,6 +1,7 @@
 import { createInstance } from 'axios-test-instance';
 import FormData from 'form-data';
 import fs from 'fs-extra';
+import { omit } from 'lodash';
 import path from 'path';
 
 import createServer from '../utils/createServer';
@@ -43,63 +44,28 @@ afterAll(async () => {
   await db.close();
 });
 
-describe('createBlockDefinition', () => {
-  it('should be possible to register a block definition', async () => {
-    const { data } = await instance.post(
-      '/api/blocks',
-      {
-        id: '@xkcd/test',
-        description: 'This block has been uploaded for the purpose of unit testing.',
-      },
-      headers,
-    );
-
-    expect(data).toStrictEqual({
-      id: '@xkcd/test',
-      description: 'This block has been uploaded for the purpose of unit testing.',
-    });
-  });
-
-  it('should not be possible to register the same block definition twice', async () => {
-    await instance.post(
-      '/api/blocks',
-      {
-        id: '@xkcd/test',
-        description: 'This block has been uploaded for the purpose of unit testing.',
-      },
-      headers,
-    );
-
-    const { data } = await instance.post(
-      '/api/blocks',
-      {
-        id: '@xkcd/test',
-        description: 'This block has been uploaded for the purpose of unit testing.',
-      },
-      headers,
-    );
-
-    expect(data).toStrictEqual({
-      error: 'Conflict',
-      message: 'Another block definition with id “@xkcd/test” already exists',
-      statusCode: 409,
-    });
-  });
-});
-
-describe('getBlockDefinition', () => {
+describe('getBlock', () => {
   it('should be possible to retrieve a block definition', async () => {
-    const { data: original } = await instance.post(
-      '/api/blocks',
-      {
-        id: '@xkcd/test',
-        description: 'This block has been uploaded for the purpose of unit testing.',
-      },
-      headers,
+    const formData = new FormData();
+    formData.append(
+      'data',
+      JSON.stringify({ name: '@xkcd/test', description: 'foo', version: '1.32.9' }),
     );
+    formData.append(
+      'build/standing.png',
+      fs.createReadStream(path.join(__dirname, '__fixtures__/standing.png')),
+    );
+    formData.append(
+      'build/testblock.js',
+      fs.createReadStream(path.join(__dirname, '__fixtures__/testblock.js')),
+    );
+
+    const { data: original } = await instance.post('/api/blocks', formData, {
+      headers: { ...headers.headers, ...formData.getHeaders() },
+    });
 
     const { data: retrieved } = await instance.get('/api/blocks/@xkcd/test');
-    expect(retrieved).toStrictEqual(original);
+    expect(retrieved).toStrictEqual(omit(original, ['files']));
   });
 
   it('should return a 404 if the requested block definition doesn’t exist', async () => {
@@ -112,37 +78,53 @@ describe('getBlockDefinition', () => {
   });
 });
 
-describe('queryBlockDefinitions', () => {
+describe('queryBlocks', () => {
   it('should be possible to query block definitions', async () => {
-    const { data: apple } = await instance.post(
-      '/api/blocks',
-      {
-        id: '@xkcd/apple',
+    const formDataA = new FormData();
+    formDataA.append(
+      'data',
+      JSON.stringify({
+        name: '@xkcd/apple',
+        version: '0.0.0',
         description: 'I’ve got an apple.',
-      },
-      headers,
+      }),
+    );
+    formDataA.append(
+      'build/standing.png',
+      fs.createReadStream(path.join(__dirname, '__fixtures__/standing.png')),
     );
 
-    const { data: pen } = await instance.post(
-      '/api/blocks',
-      {
-        id: '@xkcd/pen',
+    const { data: apple } = await instance.post('/api/blocks', formDataA, {
+      headers: { ...headers.headers, ...formDataA.getHeaders() },
+    });
+
+    const formDataB = new FormData();
+    formDataB.append(
+      'data',
+      JSON.stringify({
+        name: '@xkcd/pen',
+        version: '0.0.0',
         description: 'I’ve got a pen.',
-      },
-      headers,
+      }),
     );
+    formDataB.append(
+      'build/standing.png',
+      fs.createReadStream(path.join(__dirname, '__fixtures__/standing.png')),
+    );
+
+    const { data: pen } = await instance.post('/api/blocks', formDataB, {
+      headers: { ...headers.headers, ...formDataB.getHeaders() },
+    });
 
     const { data: bam } = await instance.get('/api/blocks');
-    expect(bam).toStrictEqual([apple, pen]);
+    expect(bam).toMatchObject([omit(apple, ['files']), omit(pen, ['files'])]);
   });
 });
 
-describe('createBlockVersion', () => {
+describe('publishBlock', () => {
   it('should be possible to upload block versions where data is sent as the first parameter', async () => {
-    await instance.post('/api/blocks', { id: '@xkcd/standing' }, headers);
-
     const formData = new FormData();
-    formData.append('data', JSON.stringify({ version: '1.32.9' }));
+    formData.append('data', JSON.stringify({ name: '@xkcd/standing', version: '1.32.9' }));
     formData.append(
       'build/standing.png',
       fs.createReadStream(path.join(__dirname, '__fixtures__/standing.png')),
@@ -152,7 +134,7 @@ describe('createBlockVersion', () => {
       fs.createReadStream(path.join(__dirname, '__fixtures__/testblock.js')),
     );
 
-    const { data, status } = await instance.post('/api/blocks/@xkcd/standing/versions', formData, {
+    const { data, status } = await instance.post('/api/blocks', formData, {
       headers: { ...headers.headers, ...formData.getHeaders() },
     });
 
@@ -165,16 +147,22 @@ describe('createBlockVersion', () => {
       resources: null,
       parameters: null,
       version: '1.32.9',
+      description: null,
     });
 
     expect(status).toBe(201);
   });
 
   it('should not accept invalid action names', async () => {
-    await instance.post('/api/blocks', { id: '@xkcd/standing' }, headers);
-
     const formData = new FormData();
-    formData.append('data', JSON.stringify({ actions: { $any: {}, $foo: {} }, version: '1.32.9' }));
+    formData.append(
+      'data',
+      JSON.stringify({
+        name: '@xkcd/standing',
+        actions: { $any: {}, $foo: {} },
+        version: '1.32.9',
+      }),
+    );
     formData.append(
       'build/standing.png',
       fs.createReadStream(path.join(__dirname, '__fixtures__/standing.png')),
@@ -183,7 +171,7 @@ describe('createBlockVersion', () => {
       'build/testblock.js',
       fs.createReadStream(path.join(__dirname, '__fixtures__/testblock.js')),
     );
-    const response = await instance.post('/api/blocks/@xkcd/standing/versions', formData, {
+    const response = await instance.post('/api/blocks', formData, {
       headers: { ...headers.headers, ...formData.getHeaders() },
     });
 
@@ -194,10 +182,8 @@ describe('createBlockVersion', () => {
   });
 
   it('should be possible to upload block versions where data is sent as the last parameter', async () => {
-    await instance.post('/api/blocks', { id: '@xkcd/standing' }, headers);
-
     const formData = new FormData();
-    formData.append('data', JSON.stringify({ version: '1.32.9' }));
+    formData.append('data', JSON.stringify({ name: '@xkcd/standing', version: '1.32.9' }));
     formData.append(
       'build/standing.png',
       fs.createReadStream(path.join(__dirname, '__fixtures__/standing.png')),
@@ -207,7 +193,7 @@ describe('createBlockVersion', () => {
       fs.createReadStream(path.join(__dirname, '__fixtures__/testblock.js')),
     );
 
-    const { data, status } = await instance.post('/api/blocks/@xkcd/standing/versions', formData, {
+    const { data, status } = await instance.post('/api/blocks', formData, {
       headers: { ...headers.headers, ...formData.getHeaders() },
     });
 
@@ -220,22 +206,21 @@ describe('createBlockVersion', () => {
       layout: null,
       resources: null,
       version: '1.32.9',
+      description: null,
     });
     expect(status).toBe(201);
   });
 
   it('should not be possible to register the same block version twice', async () => {
-    await instance.post(
-      '/api/blocks',
-      {
-        id: '@xkcd/standing',
-        description: 'This block has been uploaded for the purpose of unit testing.',
-      },
-      headers,
-    );
-
     const formData = new FormData();
-    formData.append('data', JSON.stringify({ version: '1.32.9' }));
+    formData.append(
+      'data',
+      JSON.stringify({
+        name: '@xkcd/standing',
+        description: 'This block has been uploaded for the purpose of unit testing.',
+        version: '1.32.9',
+      }),
+    );
     formData.append(
       'build/standing.png',
       fs.createReadStream(path.join(__dirname, '__fixtures__/standing.png')),
@@ -245,12 +230,19 @@ describe('createBlockVersion', () => {
       fs.createReadStream(path.join(__dirname, '__fixtures__/testblock.js')),
     );
 
-    await instance.post('/api/blocks/@xkcd/standing/versions', formData, {
+    await instance.post('/api/blocks', formData, {
       headers: { ...headers.headers, ...formData.getHeaders() },
     });
 
     const formData2 = new FormData();
-    formData2.append('data', JSON.stringify({ version: '1.32.9' }));
+    formData2.append(
+      'data',
+      JSON.stringify({
+        name: '@xkcd/standing',
+        description: 'This block has been uploaded for the purpose of unit testing.',
+        version: '1.32.9',
+      }),
+    );
     formData2.append(
       'build/standing.png',
       fs.createReadStream(path.join(__dirname, '__fixtures__/standing.png')),
@@ -260,23 +252,23 @@ describe('createBlockVersion', () => {
       fs.createReadStream(path.join(__dirname, '__fixtures__/testblock.js')),
     );
 
-    const { data } = await instance.post('/api/blocks/@xkcd/standing/versions', formData2, {
+    const { data } = await instance.post('/api/blocks', formData2, {
       headers: { ...headers.headers, ...formData2.getHeaders() },
     });
 
     expect(data).toStrictEqual({
-      error: 'Conflict',
-      message: 'Block version “@xkcd/standing@1.32.9” already exists',
-      statusCode: 409,
+      error: 'Bad Request',
+      message:
+        'Version 1.32.9 is equal to or lower than the already existing @xkcd/standing@1.32.9.',
+      statusCode: 400,
     });
   });
 
   it('should require at least one file', async () => {
-    await instance.post('/api/blocks', { id: '@xkcd/standing' }, headers);
     const formData = new FormData();
-    formData.append('data', JSON.stringify({ version: '1.32.9' }));
+    formData.append('data', JSON.stringify({ name: '@xkcd/standing', version: '1.32.9' }));
 
-    const { data, status } = await instance.post('/api/blocks/@xkcd/standing/versions', formData, {
+    const { data, status } = await instance.post('/api/blocks', formData, {
       headers: { ...headers.headers, ...formData.getHeaders() },
     });
 
@@ -291,10 +283,8 @@ describe('createBlockVersion', () => {
 
 describe('getBlockVersion', () => {
   it('should be possible to retrieve a block version', async () => {
-    await instance.post('/api/blocks', { id: '@xkcd/standing' }, headers);
-
     const formData = new FormData();
-    formData.append('data', JSON.stringify({ version: '1.32.9' }));
+    formData.append('data', JSON.stringify({ name: '@xkcd/standing', version: '1.32.9' }));
     formData.append(
       'build/standing.png',
       fs.createReadStream(path.join(__dirname, '__fixtures__/standing.png')),
@@ -304,7 +294,7 @@ describe('getBlockVersion', () => {
       fs.createReadStream(path.join(__dirname, '__fixtures__/testblock.js')),
     );
 
-    const { data: created } = await instance.post('/api/blocks/@xkcd/standing/versions', formData, {
+    const { data: created } = await instance.post('/api/blocks', formData, {
       headers: { ...headers.headers, ...formData.getHeaders() },
     });
 
@@ -329,10 +319,11 @@ describe('getBlockVersion', () => {
 
 describe('getBlockVersions', () => {
   it('should be possible to fetch uploaded block versions', async () => {
-    await instance.post('/api/blocks', { id: '@xkcd/standing' }, headers);
-
     const formData = new FormData();
-    formData.append('data', JSON.stringify({ version: '1.32.9' }));
+    formData.append(
+      'data',
+      JSON.stringify({ name: '@xkcd/standing', description: 'Version 1.32.9!', version: '1.32.9' }),
+    );
     formData.append(
       'build/standing.png',
       fs.createReadStream(path.join(__dirname, '__fixtures__/standing.png')),
@@ -341,13 +332,22 @@ describe('getBlockVersions', () => {
       'build/testblock.js',
       fs.createReadStream(path.join(__dirname, '__fixtures__/testblock.js')),
     );
-    await instance.post('/api/blocks/@xkcd/standing/versions', formData, {
+    await instance.post('/api/blocks', formData, {
       headers: { ...headers.headers, ...formData.getHeaders() },
     });
 
     const { data } = await instance.get('/api/blocks/@xkcd/standing/versions');
     expect(data).toStrictEqual([
-      { actions: null, events: null, layout: null, resources: null, version: '1.32.9' },
+      {
+        name: '@xkcd/standing',
+        description: 'Version 1.32.9!',
+        actions: null,
+        events: null,
+        layout: null,
+        parameters: null,
+        resources: null,
+        version: '1.32.9',
+      },
     ]);
   });
 
@@ -356,7 +356,7 @@ describe('getBlockVersions', () => {
     expect(data).toStrictEqual({
       statusCode: 404,
       error: 'Not Found',
-      message: 'Block definition not found',
+      message: 'Block not found.',
     });
   });
 });