+++ packages/server/src/models/BlockVersion.js
@@ -4,8 +4,15 @@ export default (sequelize) => {
   const BlockVersion = sequelize.define(
     'BlockVersion',
     {
+      OrganizationId: {
+        type: DataTypes.STRING,
+        primaryKey: true,
+        allowNull: false,
+        references: { model: 'Organization' },
+      },
       name: { type: DataTypes.STRING, primaryKey: true, unique: 'blockVersionComposite' },
       version: { type: DataTypes.STRING, primaryKey: true, unique: 'blockVersionComposite' },
+      description: { type: DataTypes.TEXT },
       layout: { type: DataTypes.STRING },
       actions: { type: DataTypes.JSON },
       parameters: { type: DataTypes.JSON },
@@ -14,16 +21,15 @@ export default (sequelize) => {
     },
     {
       freezeTableName: true,
-      paranoid: true,
       createdAt: 'created',
-      updatedAt: 'updated',
-      deletedAt: 'deleted',
+      updatedAt: false,
     },
   );
 
-  BlockVersion.associate = ({ BlockAsset }) => {
+  BlockVersion.associate = ({ BlockAsset, Organization }) => {
     BlockVersion.hasMany(BlockAsset, { foreignKey: 'name', sourceKey: 'name' });
     BlockVersion.hasMany(BlockAsset, { foreignKey: 'version', sourceKey: 'version' });
+    BlockVersion.belongsTo(Organization, { foreignKey: { allowNull: false } });
   };
 
   return BlockVersion;