+++ packages/utils/src/api/components/requestBodies/index.ts
@@ -1,5 +1,4 @@
-import blockDefinition from './blockDefinition';
 import resource from './resource';
 import user from './user';
 
-export default { blockDefinition, resource, user };
+export default { resource, user };